Update delete-essay - relates #83
* Update delete-essay to make sure only the authenticated
user can delete the current essay, checking user_id.
* Updated DeleteEssay in model.js to make sure only the authenticated
user can delete the current essay, checking user_id.

+++ database/model.js
@@ -100,9 +100,10 @@ export function createNewEssay(user_id, question) {
   });
 }
 
-export function deleteCurrEssay(essayId) {
-  const DELETE_ESSAY = `DELETE FROM essays WHERE id = $1`;
-  return db.query(DELETE_ESSAY, [essayId]);
+// Deletes current essay
+export function deleteCurrEssay(essayId, user_id) {
+  const DELETE_ESSAY = `DELETE FROM essays WHERE id = $1 AND user_id = $2`;
+  return db.query(DELETE_ESSAY, [essayId, user_id]);
 }
 
 export function getEssayInfo(essayId) {
